requirements.txt
networkx>=3.2.1
matplotlib>=3.8.2
pyvis>=0.3.2
PyYAML>=6.0.1
scipy>=1.11.0
